baby steps toward latest jupyter
