+++ kbase-extension/kbase_templates/page.html
@@ -56,6 +56,9 @@
     </script>
 
     <script src="{{static_url("components/es6-promise/promise.min.js")}}" type="text/javascript" charset="utf-8"></script>
+    <script src="{{static_url('components/preact/index.js')}}" type="text/javascript"></script>
+    <script src="{{static_url('components/proptypes/index.js')}}" type="text/javascript"></script>
+    <script src="{{static_url('components/preact-compat/index.js')}}" type="text/javascript"></script>
     <script src="{{static_url("components/requirejs/require.js") }}" type="text/javascript" charset="utf-8"></script>
     <script>
       require.config({
@@ -78,6 +81,7 @@
             codemirror: 'components/codemirror',
             termjs: 'components/term.js/src/term',
             typeahead: 'components/jquery-typeahead/dist/jquery.typeahead',
+            jed: 'components/jed/jed',
             custom: '{{static_url("custom/custom")}}',
           },
           shim: {
@@ -111,6 +115,19 @@
               }
           }
       });
+
+      define("custom", function (require, exports, module) {
+          try {
+              var custom = require('custom/custom');
+              console.debug('loaded custom.js');
+              return custom;
+          } catch (e) {
+              console.error("error loading custom.js", e);
+              return {};
+          }
+      });
+
+      document.nbjs_translations = {{ nbjs_translations|safe }};
     </script>
 
     {% block meta %}
