+++ scripts/install_narrative.sh
@@ -33,13 +33,13 @@
 #
 # 8. Done!
 
-IPYTHON_VERSION=5.3.0
+IPYTHON_VERSION=5.5.0
 
 JUPYTER_NOTEBOOK_INSTALL_DIR=jupyter_notebook
 JUPYTER_NOTEBOOK_REPO=https://github.com/jupyter/notebook
-JUPYTER_NOTEBOOK_TAG=4.4.1
+JUPYTER_NOTEBOOK_TAG=5.3.0
 
-IPYWIDGETS_VERSION=6.0.0
+IPYWIDGETS_VERSION=7.1.2
 
 PYTHON=python2.7
