+++ src/requirements.txt
@@ -2,7 +2,7 @@ nose>=1.1.2
 pyzmq>=2.2
 python-qt
 pygments>=1.6
-tornado==4.5.3
+tornado==5.0.0
 pycrypto>=2.6
 rsa>=3.1.2
 pyasn1>=0.1.7
